RFC 2466: ignore deprecated candidates for new ICE candidates
As described in RFC 2462, an ipv6 address can have several states and
some should be ignored when initializing an ICE session:

In fact DEPRECATED address SHOULD NOT be used in new communications.
Also, a TENTATIVE address is not ready to be used too. So, this
patch ignores this kind of addresses when initializing a new connection.